requirements.txt
annotated-types==0.7.0
anyio==4.5.2
Authlib==1.3.2
certifi==2025.6.15
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
exceptiongroup==1.3.0
fastapi==0.115.14
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
radon==6.0.1

